renamed style files, removed unnecesary includes, moved floor(Vector), added style map to style files
