+++ SDEV230_Project.vcxproj
@@ -149,14 +149,14 @@
     <None Include="SFML\SFML.md" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="SFML\StyleLoaders.cpp" />
+    <ClCompile Include="SFML\Styles.cpp" />
     <ClCompile Include="SFML\UI_Menu.cpp" />
     <ClCompile Include="SFML\UserInterface.cpp" />
     <ClCompile Include="SFML\SFML_test.cpp" />
     <ClCompile Include="Source.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="SFML\StyleLoaders.h" />
+    <ClInclude Include="SFML\Styles.h" />
     <ClInclude Include="SFML\UI_Menu.h" />
     <ClInclude Include="SFML\UserInterface.h" />
     <ClInclude Include="SFML\SFML_test.h" />
