+++ SDEV230_Project.vcxproj.filters
@@ -33,7 +33,7 @@
     <ClCompile Include="SFML\UI_Menu.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="SFML\StyleLoaders.cpp">
+    <ClCompile Include="SFML\Styles.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
@@ -47,7 +47,7 @@
     <ClInclude Include="SFML\UI_Menu.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="SFML\StyleLoaders.h">
+    <ClInclude Include="SFML\Styles.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
