--- SFML/StyleLoaders.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void StyleLoader();
